Add/update config for smoother tests
Four main updates:

- Add `spring-commands-rspec` to make spring work with rspec:
  This means that tests load significantly faster when you don't make
  changes to the app configuration or database.

- Configure `DatabaseCleaner`
  My tests were behaving unpredictably because of the dirty database
  state, so it was time to set up DatabaseCleaner. This (should) ensure
  that each test gets a fresh slate. Config from Avdi's blog:
    www.virtuouscode.com/2012/08/31/configuring-database_cleaner-\
    with-rails-rspec-capybara-and-selenium/

- Move helper methods from `support/omniauth` to `support/helpers/`:
  It's still included in the former (since they will be used together),
  but I hope this directory structure will make more sense to others.

- Misc. rspec configuration changes:
  These include randomized test order, the option to set a seed to
  repeat test order, ability to focus on tests with `:focus`, and other
  small updates.

Interesting Note:

  - Issues with DatabaseCleaner and rspec `around` blocks:
    I was having a lot of trouble with DatabaseCleaner and the rspec
    `around` blocks–I got a lot of uniqueness constraint violation
    errors when trying to set up tests for site_manager/admin users.
    Instead of figuring out why, I just switched to `before` + `after`
    blocks, and that solved it.
