+++ Gemfile
@@ -54,14 +54,16 @@ end
 group :development do
   # Access an IRB console on exception pages or by using <%= console %> anywhere in the code.
   gem 'web-console', '>= 3.3.0'
+  # Annotate model/model spec/factory files with the database schema
+  gem 'annotate', '~> 2.7'
   # The Listen gem listens to file modifications and notifies you about the changes. https://rubygems.org/gems/listen
   gem 'listen', '>= 3.0.5', '< 3.2'
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
   gem 'spring'
   # This gem makes Spring watch the filesystem for changes using Listen rather than by polling the filesystem.
   gem 'spring-watcher-listen', '~> 2.0.0'
-  # Annotate model/model spec/factory files with the database schema
-  gem 'annotate', '~> 2.7'
+  # Implements the `rspec` command for Spring, allowing for faster test loading
+  gem 'spring-commands-rspec', '~> 1.0'
 end
 
 group :test do
